@@ -6,51 +6,53 @@
   "scripts": {
     "start": "node index",
     "test": "NODE_ENV=test jest",
-    "coverage": "NODE_ENV=test jest --coverage",
     "dist": "NODE_ENV=production webpack -p",
     "postinstall": "npm run dist"
   },
   "engines": {
-    "node": "0.10",
-    "npm": "2.10"
+    "node": "4.2.6",
+    "npm": "3.10.5"
   },
   "dependencies": {
-    "classnames": "^2.1.2",
-    "compression": "1.4.4",
-    "ejs": "2.3.1",
-    "express": "4.12.4",
-    "express-ejs-layouts": "1.1.0",
+    "classnames": "2.2.5",
+    "compression": "1.6.2",
+    "ejs": "2.4.2",
+    "express": "4.14.0",
+    "express-ejs-layouts": "2.2.0",
     "fastclick": "1.0.6",
-    "object-assign": "2.0.0",
-    "react": "0.13.3",
-    "react-router": "0.13.3",
-    "webpack": "1.9.10"
+    "object-assign": "4.1.0",
+    "react": "15.2.1",
+    "react-dom": "15.2.1",
+    "react-router": "2.6.0",
+    "webpack": "1.13.1"
   },
   "devDependencies": {
-    "assets-webpack-plugin": "0.3.1",
-    "autoprefixer-loader": "2.0.0",
-    "babel": "5.4.7",
-    "babel-core": "5.4.7",
-    "babel-eslint": "3.1.10",
-    "babel-jest": "5.2.0",
-    "babel-loader": "5.1.3",
-    "babel-runtime": "5.4.7",
-    "coveralls": "2.11.2",
-    "css-loader": "0.14.4",
-    "eslint": "0.21.2",
-    "eslint-loader": "0.11.2",
-    "eslint-plugin-react": "2.3.0",
-    "jest": "0.1.40",
-    "jest-cli": "0.4.5",
+    "assets-webpack-plugin": "3.4.0",
+    "autoprefixer-loader": "3.2.0",
+    "babel": "6.5.2",
+    "babel-core": "6.11.4",
+    "babel-eslint": "6.1.2",
+    "babel-jest": "13.2.2",
+    "babel-loader": "6.2.4",
+    "babel-preset-es2015": "6.9.0",
+    "babel-preset-react": "6.11.1",
+    "babel-runtime": "6.9.2",
+    "css-loader": "0.23.1",
+    "eslint": "3.1.1",
+    "eslint-loader": "1.4.1",
+    "eslint-plugin-react": "5.2.2",
+    "jest": "13.2.3",
+    "jest-cli": "13.2.3",
     "jsx-loader": "0.13.2",
-    "node-libs-browser": "0.5.2",
-    "react-hot-loader": "1.2.7",
+    "node-libs-browser": "1.0.0",
+    "node-sass": "3.8.0",
+    "react-hot-loader": "1.3.0",
     "react-tools": "0.13.3",
-    "sass-loader": "1.0.2",
-    "style-loader": "0.12.3",
-    "webpack-dev-server": "1.9.0",
-    "file-loader": "^0.8.4",
-    "url-loader": "^0.5.6"
+    "sass-loader": "4.0.0",
+    "style-loader": "0.13.1",
+    "webpack-dev-server": "1.14.1",
+    "file-loader": "0.9.0",
+    "url-loader": "0.5.7"
   },
   "jest": {
     "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
