@@ -1,11 +1,14 @@
 import "styles/main.sass";
-
 import React from "react";
-import Router from "react-router";
+import {
+  Router,
+  IndexRoute,
+  Route,
+  browserHistory
+} from "react-router";
+import ReactDOM from "react-dom";
 import FastClick from "fastclick";
-
 import Navigation from "components/navigation/Navigation.jsx";
-
 import Home from "components/route-handlers/Home.jsx";
 import AcademicWork from "components/route-handlers/AcademicWork.jsx";
 import WorkWork from "components/route-handlers/WorkWork.jsx";
@@ -14,8 +17,6 @@ import Contact from "components/route-handlers/Contact.jsx";
 import Resume from "components/route-handlers/Resume.jsx";
 
 import NotFound from "components/NotFound/NotFoundComponent.jsx";
-
-const {DefaultRoute, Route, RouteHandler, NotFoundRoute} = Router;
 
 const App = React.createClass({
   getInitialState() {
@@ -45,12 +46,12 @@ const App = React.createClass({
   },
 
   render() {
-    const {windowWidth} = this.state;
+    const { windowWidth } = this.state;
 
     return (
       <div className="container">
-        <Navigation links={links} />
-        <RouteHandler windowWidth={windowWidth} />
+        <Navigation links={links} location={this.props.location} />
+        {React.cloneElement(this.props.children, { windowWidth })}
       </div>
     );
   }
@@ -77,20 +78,20 @@ const links = [
 ];
 
 const routes = (
-  <Route name="app" path="/" handler={App}>
-    <DefaultRoute handler={Home} />
-    <Route name="academic-work" path="/academic-work" handler={AcademicWork} />
-    <Route name="work" path="/work" handler={WorkWork} />
-    <Route name="miscellany" path="/miscellany" handler={Miscellany} />
-    <Route name="contact" path="/contact" handler={Contact} />
-    <Route name="resume" path="/resume" handler={Resume} />
-    <NotFoundRoute handler={NotFound} />
+  <Route path="/" component={App}>
+    <IndexRoute component={Home} />
+    <Route path="/academic-work" component={AcademicWork} />
+    <Route path="/work" component={WorkWork} />
+    <Route path="/miscellany" component={Miscellany} />
+    <Route path="/contact" component={Contact} />
+    <Route path="/resume" component={Resume} />
+    <Route path="*" component={NotFound} />
   </Route>
 );
 
-React.initializeTouchEvents(true);
-
-Router.run(routes, Router.HistoryLocation, Handler => {
-  React.render(<Handler/>, document.body);
-});
+ReactDOM.render((
+  <Router history={browserHistory}>
+    {routes}
+  </Router>
+), document.getElementById("app"));
 
