@@ -1,3 +1,1 @@
-export default {
-  constrain: (value, min, max) => Math.min(Math.max(value, min), max)
-};
+export const constrain = (value, min, max) => Math.min(Math.max(value, min), max);
