@@ -1,15 +1,15 @@
 var fs = require('fs');
 var path = require('path');
-
 var express = require('express');
 var app = express();
-
 var compress = require('compression');
 var layouts = require('express-ejs-layouts');
 
-app.set('layout');
+var port = Number(process.env['PORT'] || 3001);
+var devPort = 2992;
+
 app.set('view engine', 'ejs');
-app.set('view options', {layout:'layout'});
+app.set('layout', 'index');
 app.set('views', path.join(process.cwd(), '/server/views'));
 
 app.use(compress());
@@ -18,49 +18,39 @@ app.use("/client", express.static(path.join(process.cwd(), '/client')));
 
 app.disable('x-powered-by');
 
-var env = {
-  production: process.env['NODE_ENV'] === 'production',
-  assets: JSON.parse(fs.readFileSync(path.join(process.cwd(), 'assets.json')))
-};
-
 app.get('/*', function(req, res) {
   res.render('index', {
-    env: env
+    env: {
+      production: process.env['NODE_ENV'] === 'production',
+      assets: JSON.parse(fs.readFileSync(path.join(process.cwd(), 'assets.json'))),
+      devPort: devPort
+    }
   });
 });
 
-var port = Number(process.env['PORT'] || 3001);
 app.listen(port, function () {
-  console.log('server running at localhost:3001, go refresh and see magic');
+  console.log('server running at localhost:' + port + ', go refresh and see magic');
 });
 
-if (env.production === false) {
+if (process.env['NODE_ENV'] !== 'production') {
   var webpack = require('webpack');
   var WebpackDevServer = require('webpack-dev-server');
   var webpackConfig = require('./webpack.dev.config');
 
   new WebpackDevServer(webpack(webpackConfig), {
     publicPath: '/client/',
-
     contentBase: './client/',
-
     inline: true,
-
     hot: true,
-
     stats: false,
-
     historyApiFallback: true,
-
     headers: {
-      'Access-Control-Allow-Origin': 'http://localhost:3001',
+      'Access-Control-Allow-Origin': 'http://localhost:' + port,
       'Access-Control-Allow-Headers': 'X-Requested-With'
     }
-  }).listen(3000, 'localhost', function (err, result) {
-    if (err) {
-      console.log(err);
-    }
+  }).listen(devPort, 'localhost', function (err, result) {
+    if (err) { console.log(err); }
 
-    console.log('webpack dev server listening on localhost:3000');
+    console.log('webpack dev server listening on localhost:' + devPort);
   });
 }
