@@ -5,7 +5,7 @@ var config = require('./webpack.base.config.js');
 
 if (process.env['NODE_ENV'] !== 'test') {
   config.entry = [
-    'webpack-dev-server/client?http://localhost:3000',
+    'webpack-dev-server/client?http://localhost:2992',
     'webpack/hot/dev-server',
     config.entry
   ];
@@ -18,7 +18,11 @@ config.plugins = config.plugins.concat([
 ]);
 
 config.module.loaders = config.module.loaders.concat([
-  {test: /\.jsx?$/, loaders: [ 'react-hot', 'babel'], exclude: /node_modules/}
+  {
+    test: /\.jsx?$/,
+    loaders: ['react-hot', 'babel?presets[]=es2015,presets[]=react'],
+    exclude: /node_modules/
+  }
 ]);
 
 module.exports = config;
