@@ -30,7 +30,14 @@ config.plugins = config.plugins.concat([
 ]);
 
 config.module.loaders = config.module.loaders.concat([
-  {test: /\.jsx?$/, loaders: [ 'babel'], exclude: /node_modules/}
+  {
+    test: /\.jsx?$/,
+    loader: 'babel',
+    query: {
+      presets: ['es2015', 'react']
+    },
+    exclude: /node_modules/
+  }
 ]);
 
 module.exports = config;
